requirements.txt
kafka-python
requests
pyspark
fastapi
uvicorn
psycopg2-binary